breakup confteset and rearrange folder
