+++ tests/_utilities/fixtures/vcr.py
@@ -1,0 +1,10 @@
+import pytest
+
+
+@pytest.fixture(scope="session")
+def vcr_config():
+    """Overwrite headers and query parameters where key can be leaked"""
+    return {
+        "filter_headers": [("authorization", "DUMMY")],
+        "filter_query_parameters": ["key"],
+    }
